processor/requirements.txt
# UDP Receiver の依存関係
# AIテキスト編集機能のための依存関係（Gemini API使用）
google-generativeai>=0.3.0
requests>=2.25.0
